Update documentation to reflect build/push and v3 changes
Adds a user guide on how to use the build/push functionality of Docker
images to the user guide.

Adds v3 documentation as well as cleans up the user guide as well.

I also do minor changes such as removing `$ ` in the CLI examples, as
it's easier to copy and paste without the `$ ` there (when double
clicking).

I've also changed the `console` yaml indicitors to `sh` since it doesn't
matter / it's aliased, see:
https://github.com/github/linguist/blob/master/lib/linguist/languages.yml