Updates the conversion document to reflect changes
This updates the conversion document to outline that we do not support
minor versions since libcompose does not support 2.1 or 2.2 as well as
constant changes to docker/cli stack code regarding 3.1 and above.

I also update the conversion document to better reflect our support on
version 3 as well as update current values.