Add support for push-image with --build local (#1257)
The PR https://github.com/kubernetes/kompose/pull/1147 introduced a
push-image parameter for the 'up' command that allows to disable
pushing the image to the registry. The PR changed the behavior of the
'convert' command that now no longer pushes the image.

This commit introduces the same parameter for the 'convert' command
and reintroduces the original behavior of pushing the image.